Implemented bounding box drawing
The bounding box drawing is still not working properly but we can draw a rectangle over a selected area at a selected frame.

I noticed that we can't draw on top of ImageIcon so I was thinking of implementing a "Draw button" that will open up another frame/window containing the BufferedImage at the specified slider location, draw on top of that BufferedImage, then updating/setting this image to an ImageIcon for display on the main picture JFrame. I believe this should be able to get rid of the errors that pop up when selecting rectangles on different frames.